refactor http code to rest_api.py and TLS code to tls.py
